🧹 chore: Rename configResolution.ts to resolution.ts
