+++ packages/data-schemas/src/app/index.ts
@@ -5,4 +5,4 @@ export * from './specs';
 export * from './turnstile';
 export * from './vertex';
 export * from './web';
-export * from './configResolution';
+export * from './resolution';
